requirements.txt
absl-py==0.8.0
actionlib==1.11.13
alabaster==0.7.12
albumentations==0.4.5
align==0.0.11
anaconda-client==1.7.2
anaconda-navigator==1.8.7
angles==1.9.12
appdirs==1.4.3
asn1crypto==1.0.1
astor==0.8.0
astroid==2.3.1
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bleach==1.5.0
bondpy==1.8.3
boto3==1.12.36
botocore==1.15.36
cachetools==4.0.0
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.20
catkin-pkg==0.4.14
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
conda==4.7.12
conda-build==3.10.5
conda-package-handling==1.6.0
conda-verify==3.4.2
cryptography==2.7
cv-bridge==1.12.8
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
distlib==0.3.0
docutils==0.15.2
dynamic-reconfigure==1.5.50
empy==3.3.4
entrypoints==0.3
filelock==3.0.12
future==0.17.1
gast==0.3.2
gazebo-plugins==2.5.19
gazebo-ros==2.5.19
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
gensim==3.8.1
glob2==0.7
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-pasta==0.1.7
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.16.1
h5py==2.9.0
html5lib==0.9999999
idna==2.8
image-geometry==1.12.8
imageio==2.6.1
imagesize==1.1.0
imgaug==0.2.6
importlib-metadata==1.6.0
importlib-resources==1.4.0
interactive-markers==1.11.4
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
isort==4.3.21
jedi==0.15.1
jeepney==0.4.1
Jinja2==2.10.3
jmespath==0.9.5
joblib==0.14.1
joint-state-publisher==1.12.15
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.4
jupyter-core==4.6.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.1.0
laser-geometry==1.6.5
lazy-object-proxy==1.4.2
Mako==1.0.6
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
message-filters==1.12.14
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
module==0.0.4
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numpy==1.17.2
numpydoc==0.9.1
olefile==0.46
opencv-python==4.2.0.32
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
pluginlib==1.11.3
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.9.2
psutil==5.6.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pygpu==0.6.9
pylint==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.0
python-qt-binding==0.3.7
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
qt-dotgraph==0.3.17
qt-gui==0.3.17
qt-gui-cpp==0.3.17
qt-gui-py-common==0.3.17
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
requests==2.22.0
resource-retriever==1.12.6
rope==0.14.0
rosbag==1.12.14
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosdep==0.16.1
rosdistro==0.7.5
rosgraph==1.12.14
rosinstall==0.7.8
rosinstall-generator==0.1.18
roslaunch==1.12.14
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.6
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.1.10
rospy==1.12.14
rosserial-client==0.7.7
rosserial-python==0.7.7
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.6
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.14
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rsa==4.0
ruamel-yaml==0.15.46
rviz==1.12.17
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.3.1
SecretStorage==3.1.1
Send2Trash==1.5.0
sensor-msgs==1.12.7
six==1.12.0
smach==2.0.1
smach-ros==2.0.1
smart-open==1.10.0
smclib==1.8.3
snowballstemmer==2.0.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.6
spyder-kernels==0.5.2
tensorboard==1.7.0
tensorflow==1.13.1
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
Theano==0.9.0
topic-tools==1.12.14
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.24.2
vcstools==0.1.42
virtualenv==20.0.16
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wstool==0.1.17
wurlitzer==1.0.3
xacro==1.11.3
zipp==3.1.0
